+++ src/commands/mod.rs
@@ -2,6 +2,7 @@ mod banner;
 pub mod build;
 pub mod client;
 pub mod clean;
+pub mod config;
 pub mod deploy;
 pub mod help;
 pub mod idl;